requirements.txt
alembic==1.4.3
astroid==2.4.2
attrs==20.2.0
click==7.1.2
coverage==5.3
Faker==4.14.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
